+++ src/lib/mrvGL/mrvTimelineViewport.cpp
@@ -1008,6 +1008,7 @@ namespace mrv
             if (value == p.hdrOptions)
                 return;
             p.hdrOptions.algorithm = value.algorithm;
+            p.hdrOptions.gamutMapping = value.gamutMapping;
             redrawWindows();
         }